Baselines/16-20/DeepHash-pytorch-master/BLIP2/requirement.txt
accelerate==0.24.1
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.21.0
altair==5.1.2
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.37.0
bleach==6.1.0
blis==0.7.11
braceexpand==0.1.7
brotlipy==0.7.0
catalogue==2.0.10
cchardet==2.1.7
certifi 
cffi
chardet==5.1.0
charset-normalizer
click==8.1.7
comm==0.1.4
confection==0.1.3
contourpy==1.0.7
cryptography
cycler==0.11.0
cymem==2.0.8
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.19.3
dill==0.3.7
docker-pycreds==0.4.0
easydict==1.9
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.0
fairscale==0.4.13
faiss==1.7.4
fastapi==0.104.0
fastjsonschema==2.18.1
ffmpy==0.3.1
filelock==3.9.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.40
gradio==3.24.1
gradio_client==0.0.8
h11==0.14.0
h5py==3.13.0
hf-xet==1.1.2
httpcore==0.18.0
httpx==0.25.0
huggingface==0.0.1
huggingface-hub==0.14.1
idna
imageio==2.31.5
imageio-ffmpeg==0.6.0
importlib-metadata==6.8.0
importlib-resources==5.12.0
iniconfig==2.1.0
intel-openmp==2023.2.0
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.4.0
jupyter_packaging==0.12.3
jupyter_server==2.8.0
jupyter_server_fileid==0.9.3
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
linkify-it-py==2.0.2
llvmlite==0.41.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mkl==2023.2.0
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbclassic==1.3.1
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2
nltk==3.8.1
notebook==6.5.7
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.14.1
openai==0.27.0
opencv-python==4.7.0.72
orjson==3.9.9
overrides==7.4.0
packaging==23.0
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.2
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow
platformdirs==3.11.0
pluggy==1.6.0
portalocker==2.8.2
preshed==3.0.9
proglog==0.1.12
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser 
pydantic==1.10.13
pydub==0.25.1
Pygments==2.16.1
pynndescent==0.5.10
pyOpenSSL 
pyparsing==3.0.9
PySocks 
pytest==8.4.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2022.10.31
requests
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
six
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tbb==2021.10.0
tenacity==8.2.2
terminado==0.17.1
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.13.2
toolz==0.12.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.3.3
tqdm==4.64.1
traitlets==5.11.2
transformers==4.28.0
triton==2.1.0
typer==0.7.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
umap-learn==0.5.4
uri-template==1.3.0
urllib3
uvicorn==0.23.2
wandb==0.15.12
wasabi==1.1.2
wcwidth==0.2.8
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
widgetsnbextension==4.0.14
xxhash==3.4.1
y-py==0.6.2
yarl==1.8.2
ypy-websocket==0.8.4
zipp==3.14.0
